Add historic rates (close #36) (#49)
* Add historic rates (close #36)

* Add exchange impl

* Update get_historic_rates to use CandleRequestParams

+++ coinbase/src/model/mod.rs
@@ -198,7 +198,7 @@ pub struct OrderRequest {
 pub struct CandleRequestParams {
     #[serde(flatten)]
     pub daterange: Option<DateRange>,
-    pub granularity: Option<i64>,
+    pub granularity: Option<u32>,
 }
 
 #[derive(Serialize, Deserialize, Debug)]
@@ -300,7 +300,7 @@ pub struct Paginator {
     pub after: Option<i64>,
 }
 
-#[derive(Serialize, Deserialize, Debug)]
+#[derive(Clone, Copy, Serialize, Deserialize, Debug)]
 #[serde(rename_all = "camelCase")]
 pub struct DateRange {
     #[serde(with = "opt_naive_datetime_from_string")]
